Fix/tests

+++ src/services/techs/deleteTech.services.ts
@@ -2,7 +2,7 @@ import AppDataSource from "../../data-source";
 import { Tech } from "../../entities/tech.entity";
 import { AppError } from "../../errors";
 
-const deleteTechService = async (id: string) => {
+const deleteTechService = async (id: string, userId: string) => {
   const TechRepository = AppDataSource.getRepository(Tech);
 
   const searchTech = await TechRepository.findOneBy({ id });
@@ -10,8 +10,11 @@ const deleteTechService = async (id: string) => {
   if (!searchTech) {
     throw new AppError("Tech Not Found", 404);
   } else {
-    const TechUpdate = await TechRepository
-      .createQueryBuilder()
+    if (searchTech.userId !== userId) {
+      throw new AppError("user does not have permission", 403);
+    }
+
+    const TechUpdate = await TechRepository.createQueryBuilder()
       .delete()
       .from(Tech)
       .where("id = :id", { id: id })
